# project-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Forge API
  description: |
    REST API for Forge, a file-system-backed project management tool for software teams.
    All request/response bodies use **snake_case** field names.
    Paginated list endpoints return `{ items, total, page, limit, total_pages }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/features:
    get:
      summary: List features (paginated)
      operationId: listFeatures
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: status
          in: query
          schema: { type: string }
          description: Comma-separated list of statuses (draft, deferred, ready, in-progress, review, done)
        - name: priority
          in: query
          schema: { type: string }
          description: Comma-separated list of priorities (critical, high, medium, low)
        - name: assignee
          in: query
          schema: { type: string }
          description: Filter by assignee name, or `__unassigned__` for unassigned features
        - name: points_min
          in: query
          schema: { type: number }
          description: Minimum points filter
        - name: points_max
          in: query
          schema: { type: number }
          description: Maximum points filter
        - name: tags
          in: query
          schema: { type: string }
          description: Comma-separated list of tags to filter by
        - name: sort
          in: query
          schema:
            type: string
            enum: [order, title, status, priority, modified]
            default: order
        - name: dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Paginated list of features
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFeatures"

    post:
      summary: Create a feature
      operationId: createFeature
      tags: [Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureInput"
      responses:
        "201":
          description: Feature created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/wiki:
    get:
      summary: List wiki docs (paginated)
      operationId: listWikiDocs
      tags: [Wiki]
      parameters:
        - $ref: "#/components/parameters/page"
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 100 }
          description: Items per page (default 50)
      responses:
        "200":
          description: Paginated list of wiki docs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedWikiDocs"

    post:
      summary: Create a wiki doc (auto-generated slug)
      operationId: createWikiDoc
      tags: [Wiki]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveWikiDocInput"
      responses:
        "201":
          description: Doc created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiDoc"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/wiki/{slug}:
    parameters:
      - $ref: "#/components/parameters/slug"

    get:
      summary: Get a wiki doc by slug
      operationId: getWikiDoc
      tags: [Wiki]
      responses:
        "200":
          description: Wiki doc detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiDoc"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Create or update a wiki doc at a specific slug
      operationId: saveWikiDoc
      tags: [Wiki]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveWikiDocInput"
      responses:
        "200":
          description: Doc saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiDoc"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a wiki doc
      operationId: deleteWikiDoc
      tags: [Wiki]
      responses:
        "200":
          description: Doc deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  slug: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Roadmap ───────────────────────────────────────────────────────────────
  /api/roadmap:
    get:
      summary: Get the full roadmap
      operationId: getRoadmap
      tags: [Roadmap]
      responses:
        "200":
          description: Full roadmap with cards, rows, and columns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"

    put:
      summary: Save the full roadmap (bulk update)
      operationId: saveRoadmap
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Roadmap"
      responses:
        "200":
          description: Updated roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"
...
  /api/chat:
    post:
      summary: Stream an AI chat response
      description: |
        Sends messages to an LLM with project context. Returns a streaming UIMessageStream response.
        The AI has tools to replace/edit wiki docs, generate roadmaps, and convert roadmap cards to features.
      operationId: chat
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items: { type: object }
                  description: Array of UIMessage objects from the AI SDK
                context:
                  $ref: "#/components/schemas/ChatContext"
                model:
                  type: string
                  description: "Model key (e.g. anthropic/claude-sonnet-4-5, openai/gpt-4o)"
                  default: anthropic/claude-sonnet-4-5
      responses:
        "200":
          description: Streaming UIMessageStream response
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── AI Sessions ───────────────────────────────────────────────────────────
...
components:
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
      description: Page number
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 25, minimum: 1, maximum: 100 }
      description: Items per page
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    ValidationError:
      description: Validation error (422)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, description: "Semicolon-separated list of field-level errors" }
    Conflict:
      description: Resource already exists (409)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }

  # ── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    FeatureStatus:
      type: string
      enum: [draft, deferred, ready, in-progress, review, done]
...
    Priority:
      type: string
      enum: [critical, high, medium, low]

    # ── Feature ────────────────────────────────────────────────────────────
    Feature:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/FeatureStatus" }
        priority: { $ref: "#/components/schemas/Priority" }
        assignees:
          type: array
          items: { type: string }
        points: { type: number, nullable: true }
        sprint: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        created: { type: string, description: ISO 8601 date }
        modified: { type: string, description: ISO 8601 date }
        roadmap_card: { type: string, nullable: true, description: Linked roadmap card ID }
        body: { type: string }
        order: { type: number }
    PaginatedFeatures:
      type: object
      properties:
        features:
          type: array
          items: { $ref: "#/components/schemas/Feature" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }
    CreateFeatureInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        roadmap_card_id: { type: string, description: Roadmap card to link }
...
    WikiDoc:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        updated_at: { type: string, description: ISO 8601 date }
        body: { type: string }
        order: { type: number }
        dirty: { type: boolean, description: Whether the doc has unsaved changes }
    PaginatedWikiDocs:
      type: object
      properties:
        docs:
          type: array
          items: { $ref: "#/components/schemas/WikiDoc" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }
    SaveWikiDocInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        icon: { type: string, nullable: true }
        body: { type: string, default: "" }

    # ── Roadmap ────────────────────────────────────────────────────────────
    Roadmap:
      type: object
      properties:
        title: { type: string }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items: { $ref: "#/components/schemas/RoadmapRow" }
        cards:
          type: array
          items: { $ref: "#/components/schemas/RoadmapCard" }
        modified: { type: string, nullable: true }
    RoadmapCard:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        notes: { type: string }
        column: { type: string }
        row: { type: string }
        order: { type: number }
        feature_slug: { type: string, nullable: true }
    RoadmapRow:
      type: object
      properties:
        label: { type: string }
        color: { type: string, nullable: true }
...
    ChatContext:
      type: object
      nullable: true
      properties:
        type:
          type: string
          enum: [wiki_doc, roadmap, global]
        slug: { type: string }
        title: { type: string }
        body: { type: string }

    # ── Sessions ───────────────────────────────────────────────────────────
...